# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Livia API — multi-tenant appointment operating system
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /businesses:
    post:
      operationId: createBusiness
      tags: [businesses]
      summary: Create a new business
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBusinessBody"
      responses:
        "201":
          description: Created business
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Business"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/Conflict"
...
  /public/b/{slug}/chat:
    post:
      operationId: sendPublicChatMessage
      tags: [chat]
      summary: Send a message to the public AI booking assistant
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PublicChatBody"
      responses:
        "200":
          description: Assistant reply
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicChatResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    internalOpsSecret:
      type: apiKey
      in: header
      name: X-Internal-Ops-Secret
      description: Livia Inc operator secret (INTERNAL_OPS_SECRET or INTERNAL_CRON_SECRET fallback)
...
  responses:
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    Conflict:
      description: Conflict (e.g. slug taken, slot unavailable)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    Business:
      type: object
      properties:
        id:
          type: string
        ownerId:
          type: string
        name:
          type: string
        slug:
          type: string
        description:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        addressLine1:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        country:
          type: ["string", "null"]
        timezone:
          type: string
        locale:
          type: string
          description: BCP-47 locale for customer-facing copy and formatting (e.g. en-IE)
        currency:
          type: string
          description: ISO 4217 currency code (e.g. EUR)
        logoUrl:
          type: ["string", "null"]
        coverImageUrl:
          type: ["string", "null"]
        instagramHandle:
          type: ["string", "null"]
        aiEnabled:
          type: string
        aiTone:
          type: string
        aiGreeting:
          type: ["string", "null"]
        aiKnowledge:
          type: ["string", "null"]
        aiCanBookDirectly:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        onboardingState:
          $ref: "#/components/schemas/OnboardingState"
      required: [id, ownerId, name, slug, timezone, createdAt, updatedAt]
    OnboardingState:
      type: object
      description: Self-serve wizard progress (acts A1–A12)
      properties:
        currentAct:
          $ref: "#/components/schemas/OnboardingActId"
        completedActs:
          type: array
          items:
            $ref: "#/components/schemas/OnboardingActId"
        percentComplete:
          type: integer
          minimum: 0
          maximum: 100
        checklist:
          $ref: "#/components/schemas/OnboardingChecklist"
        updatedAt:
          type: string
          format: date-time
      required: [currentAct, completedActs, percentComplete]
    OnboardingActId:
      type: string
      enum:
        - a1_create_business
        - a2_shop_profile
        - a3_service_menu
        - a4_team
        - a5_hours
        - a6_liv
        - a7_channels
        - a8_public_link
        - a9_billing
        - a10_invite_team
        - a11_migration
        - a12_go_live
    OnboardingChecklist:
      type: object
      properties:
        testBooking: { type: boolean }
        livEnabled: { type: boolean }
        publicLinkShared: { type: boolean }
        smsOrVoiceConnected: { type: boolean }
        teamInvited: { type: boolean }
        billingStarted: { type: boolean }
        servicesConfirmed: { type: boolean }
        hoursConfirmed: { type: boolean }
...
    CreateBusinessBody:
      type: object
      properties:
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        category:
          type: string
        email:
          type: string
        phone:
          type: string
        timezone:
          type: string
        city:
          type: string
        country:
          type: string
        jurisdiction:
          type: string
          description: ISO jurisdiction pack code (IE, GB, DE, ES, IT, NL, PL)
          enum: [IE, GB, DE, ES, IT, NL, PL]
        vertical:
          $ref: "#/components/schemas/BusinessVertical"
        tier:
          $ref: "#/components/schemas/BusinessTier"
        seedDefaults:
          type: boolean
          description: When true (default), seed services and staff from the vertical pack
        logoUrl:
          type: string
        instagramHandle:
          type: string
      required: [name, slug]
    BusinessVertical:
      type: string
      enum: [hair, beauty, body-art, wellness, fitness, medspa, allied-health]
    BusinessTier:
      type: string
      enum: [solo, studio, chain, chair-host, white-label]
...
    PublicChatBody:
      type: object
      properties:
        conversationId:
          type: string
        message:
          type: string
        customerName:
          type: string
        customerEmail:
          type: string
        customerPhone:
          type: string
      required: [message]
    PublicChatResponse:
      type: object
      properties:
        conversationId:
          type: string
        reply:
          type: string
        bookingId:
          type: ["string", "null"]
        status:
          type: string
          enum: [OPEN, HANDED_OFF, CLOSED]
      required: [conversationId, reply, status]
...